Refactor DataFetchers
