+++ graphql/src/main/java/com/scalar/db/graphql/datafetcher/QueryGetDataFetcher.java
@@ -9,6 +9,8 @@ import com.scalar.db.api.Get;
 import com.scalar.db.api.Result;
 import com.scalar.db.exception.storage.ExecutionException;
 import com.scalar.db.exception.transaction.CrudException;
+import com.scalar.db.graphql.schema.TableGraphQlModel;
+import com.scalar.db.io.Key;
 import graphql.execution.DataFetcherResult;
 import graphql.schema.DataFetcher;
 import graphql.schema.DataFetchingEnvironment;
@@ -20,12 +22,13 @@ import org.slf4j.LoggerFactory;
 public class QueryGetDataFetcher
     implements DataFetcher<DataFetcherResult<Map<String, Map<String, Object>>>> {
   private static final Logger LOGGER = LoggerFactory.getLogger(QueryGetDataFetcher.class);
-  private final DistributedStorage storage;
-  private final DataFetcherHelper helper;
 
-  public QueryGetDataFetcher(DistributedStorage storage, DataFetcherHelper helper) {
+  private final TableGraphQlModel tableGraphQlModel;
+  private final DistributedStorage storage;
+
+  public QueryGetDataFetcher(TableGraphQlModel tableGraphQlModel, DistributedStorage storage) {
+    this.tableGraphQlModel = tableGraphQlModel;
     this.storage = storage;
-    this.helper = helper;
   }
 
   @Override
@@ -47,10 +50,10 @@ public class QueryGetDataFetcher
                   dbResult.getValue(fieldName).ifPresent(value -> data.put(fieldName, value.get()));
                 }
               });
-      result.data(ImmutableMap.of(helper.getObjectTypeName(), data.build()));
+      result.data(ImmutableMap.of(tableGraphQlModel.getObjectType().getName(), data.build()));
     } catch (CrudException | ExecutionException e) {
       LOGGER.warn("Scalar DB get operation failed", e);
-      result.error(DataFetcherHelper.getGraphQLError(e, environment));
+      result.error(DataFetcherUtils.createGraphQLError(e, environment));
     }
 
     return result.build();
@@ -60,13 +63,13 @@ public class QueryGetDataFetcher
   @SuppressWarnings("unchecked")
   Get createGet(Map<String, Object> getInput, DataFetchingEnvironment environment) {
     Map<String, Object> key = (Map<String, Object>) getInput.get("key");
+    Key partitionKey = DataFetcherUtils.createPartitionKeyFromKeyArgument(tableGraphQlModel, key);
+    Key clusteringKey = DataFetcherUtils.createClusteringKeyFromKeyArgument(tableGraphQlModel, key);
     Get get =
-        new Get(
-                helper.createPartitionKeyFromKeyArgument(key),
-                helper.createClusteringKeyFromKeyArgument(key))
-            .forNamespace(helper.getNamespaceName())
-            .forTable(helper.getTableName())
-            .withProjections(DataFetcherHelper.getProjections(environment));
+        new Get(partitionKey, clusteringKey)
+            .forNamespace(tableGraphQlModel.getNamespaceName())
+            .forTable(tableGraphQlModel.getTableName())
+            .withProjections(DataFetcherUtils.getProjections(environment));
     String consistency = (String) getInput.get("consistency");
     if (consistency != null) {
       get.withConsistency(Consistency.valueOf(consistency));
@@ -78,13 +81,13 @@ public class QueryGetDataFetcher
   @VisibleForTesting
   Optional<Result> performGet(DataFetchingEnvironment environment, Get get)
       throws CrudException, ExecutionException {
-    DistributedTransaction transaction = DataFetcherHelper.getCurrentTransaction(environment);
+    DistributedTransaction transaction = DataFetcherUtils.getCurrentTransaction(environment);
     if (transaction != null) {
       LOGGER.debug("running Get operation with transaction: {}", get);
       return transaction.get(get);
     } else {
       LOGGER.debug("running Get operation with storage: {}", get);
-      helper.failIfConsensusCommitTransactionalTable();
+      DataFetcherUtils.failIfConsensusCommitTransactionalTable(tableGraphQlModel);
       return storage.get(get);
     }
   }
